`super` has to be treated slightly differently.
In a sub-module, `super` is inbetween `pub(crate)` and private
visibility. However, `super` isn't valid in top-level modules. If you
try using it in such cases you get an error:

  There are too many leading `super` keywords

It's not a great error but it really means "you can't use `super` (or
`super super` or ...) to reference beyond a top-level module".

This meant that depub missed depubing top-level modules properly. This
commit fixes that.